fix: HTTP method 검증 추가 refactor 과정에서 삭제된 HTTP method 검증 과정을 추가했다.

+++ src/main/java/webserver/request/service/UserLogoutHandler.java
@@ -19,6 +19,11 @@ public class UserLogoutHandler implements RequestProcessor {
         HTTPResponseBody responseBody = null;
 
         try {
+            String method = requestHeader.getMethod();
+            if (!method.equals("POST")) {
+                throw new HTTPExceptions.Error405("Method not supported " + method);
+            }
+
             for (Cookie cookie : cookieList) {
                 if (cookie.getName().equals("SESSIONID")) {
                     Database.deleteSession(cookie.getValue());
